+++ NEWS.md
@@ -4,6 +4,7 @@
 * Added the function `abundance_classes` to transform abundance-coverage data into abundance-coverage classes.
 * New macroinvertebrate-based indices! Functions `dehli` and `flow_T` added.
 * New functions for building reference datasets from online resources! Function `get_gbif_taxa_tree`, `get_iucn_taxa_tree`, `get_nbn_taxa_tree` and `get_worms_taxa_tree` added.
+* New macrophyte-based index! Function `ibmr` added.
 
 2# biomonitoR 0.9.2